TakeHomeAssignment/requirements.txt
mistralai>=1.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
pymongo>=4.0.0
pytest>=8.0.0
